+++ src/test/kotlin/dk/lldata/typescriptpoet/test/KitchenSinkJavaTests.kt
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2017 Outfox, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package dk.lldata.typescriptpoet.test
+
+import org.hamcrest.CoreMatchers.equalTo
+import org.hamcrest.MatcherAssert.assertThat
+import org.junit.jupiter.api.DisplayName
+import org.junit.jupiter.api.Test
+
+@DisplayName("Kitchen Sink, Java vs Kotlin")
+class KitchenSinkJavaTests {
+
+  @Test
+  @DisplayName("The Java kitchen sink emits the same file as the Kotlin one")
+  fun testSameFile() {
+    assertThat(KitchenSinkJava.emit(), equalTo(KitchenSink.emit()))
+  }
+}